requirements.txt

gunicorn
django-heroku
argcomplete==1.10.0
asgiref==3.2.10
astroid==2.4.2
attrs==19.3.0
autopep8==1.5.3
backcall==0.1.0
beautifulsoup4==4.8.0
bleach==3.1.5
blis==0.4.1
catalogue==1.0.0
certifi==2020.6.20
chardet==3.0.4
cloudpickle==1.5.0
cymem==2.0.3
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.8
docx2txt==0.8
EbookLib==0.17.1
entrypoints==0.3
extract-msg==0.23.1
idna==2.10
image==1.5.32
IMAPClient==2.1.0
img2pdf==0.3.6
importlib-metadata==1.7.0
ipykernel==5.2.1
ipython==7.14.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.0
Jinja2==2.11.2
joblib==0.16.0
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
lazy-object-proxy==1.4.3
lxml==4.5.1
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
numpy==1.19.0
olefile==0.46
opencv-python==4.3.0.36
packaging==20.3
pandocfilters==1.4.2
parso==0.7.0
pathlib==1.0.1
pdfminer.six==20181108
pdfplumber==0.5.21
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
plac==1.2.0
preshed==3.0.2
prometheus-client==0.7.1
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pycodestyle==2.6.0
pycryptodome==3.9.8
Pygments==2.6.1
pylint==2.5.3
pyparsing==2.4.7
pyrsistent==0.16.0
pytesseract==0.3.4
python-dateutil==2.8.1
python-pptx==0.6.18
pytz==2020.1
pyzmq==19.0.1
requests==2.24.0
scikit-learn==0.23.1
scipy==1.5.0
semver==2.10.2
Send2Trash==1.5.0
six==1.14.0
sortedcontainers==2.2.2
soupsieve==2.0.1
spacy==2.3.0
SpeechRecognition==3.8.1
sputnik==0.9.3
sqlparse==0.3.1
srsly==1.0.2
terminado==0.8.3
tesseract==0.1.3
testpath==0.4.4
textract==1.6.3
thinc==7.4.1
threadpoolctl==2.1.0
toml==0.10.1
tornado==6.0.4
tqdm==4.47.0
traitlets==4.3.3
typed-ast==1.4.1
tzlocal==1.5.1
ujson==3.0.0
unicodecsv==0.14.1
urllib3==1.25.9
Wand==0.6.1
wasabi==0.7.0
wcwidth==0.1.9
webencodings==0.5.1
wrapt==1.12.1
xlrd==1.2.0
XlsxWriter==1.2.9
zipp==3.1.0